Moved initialization responsibility from MainActivity to ApplicationState.
