+++ app/src/main/java/com/example/ttapp/MainActivity.java
@@ -6,7 +6,7 @@ import android.content.Context;
 import android.os.Bundle;
 
 import com.example.ttapp.APIRequester.APIRequester;
-import com.example.ttapp.APIRequester.DebugRequester;
+import com.example.ttapp.debug.DebugRequester;
 import com.example.ttapp.APIRequester.TTRequester;
 import com.example.ttapp.database.MongoDB;
 
@@ -16,11 +16,7 @@ public class MainActivity extends AppCompatActivity {
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_main);
-        TTRequester.initialize(
-                this.getApplicationContext(),
-                this.getPreferences(Context.MODE_PRIVATE));
-        DebugRequester.initialize(this.getApplicationContext());
-        MongoDB.initialize(this.getApplicationContext());
+        ApplicationState.initializeComponentsRequiringActivity(this);
         ApplicationState.enterState(ApplicationState.State.PRODUCTION);
         APIRequester.getInstance().sendOldResponses();
     }
